+++ src/client/mod.rs
@@ -1,5 +1,4 @@
 pub use self::{
-    builder::ConnectionBuilder,
     connection::Connection,
     connection_like::ConnectionLike,
     stream::Stream,
@@ -8,7 +7,6 @@ pub use self::{
 
 pub mod schema;
 
-mod builder;
 mod connection;
 mod connection_like;
 mod stream;